More changes for processing items on providers
The Alias processing is now complete, as per our list discsussion. Have
moved some methods around for cleanlyness. Also have refactored to have
a ProviderThread base class. This feels a little clunky, but the
alternative was quite significant code duplication.

Tests now cover the new Alias processing thread, and handle errors
on a single provider, on multiple providers, and on providers in sequence.