Now allows configuration of the database in app.cfg
This is used in both the test suite and when running as an application.
The TOTALIMPACT_CONFIG environment setting should now point to a flask
config file which only specifies override values.
Reverted last changes which used the json config. This is depricated.
Now removing the json config to avoid confusion.