Huge refactor of providers. This is to stop providers needing to deal with items
Also had made quite a bit of changes to logging. This is in flux at present and
needs completion. I've added context information to logging so that providers
can just use their own loggers and they don't need these passed as arguements.

This branch is broken, but works for one of the functional tests. The test suite
is very tight, and needs updating.