Bringing into line with the providers changes, and fixing response parsing
We were trying to use beautiful soup to read the response from wikipedia.
Much better to just use XML as it should be a valid XML response. Also
now reading from the result value field, rather than trying to enumerate
results. This was probably giving incorrect values before, when the
number of references was large.