Restoring previous test data files
These are a bit of a mish-mash in terms of naming. We have moved these to
extras, although we really should be using these inside the test suite as
well, so maybe test/data is better? Regardless, the format we have in extras
under sample_provider_pages isn't sufficient - sometimes we're going to need
alias or metrics responses from a provider for more than 1 test item.

Restoring for now, so that daily checks pass. If we are going to clean this
up we'll need to fix the issues.