adds WEBAPP_ROOT env var for use in integration tests.
This is inelegant because it means that our api knows about our webapp. It is
prettier if it does not. The alternative is to run the functional tests from
and endpoint on the webapp. This might be better, since the webapp does have
to know where the API_ROOT is, and does have to know how to query it. The
only bad thing is that most of the testing is *of* the ti-core code, and
so it is handy to have it in the ti-core repo.