add two views to couchdb for use by analytics
these two views also have reduce functions but our create-view-couchdb-docs code doesn't have any way to add the reduce function to the docs.  Not adding this now since plan is to move away from couch, and these views are only used by analytics calls.  Reduce function is documented in the bottom of the .js files as a comment.