DELETE /collection/:cid can only delete items if they're just owned by test users.
This way items that test users own aren't deleted if regular users are also using them.
Note that this can make tests less deterministic; you can't be sure that a test item will
be deleted, since a real user may have added it to a collection.

Also note that this requires a call back to the webapp to get test-user CIDs. Pretty
convoluted since this method is probably being called by the webapp in the first place;
but I think is necessary evil given current postgres-couchdb split.